moved geospatial crud/schemas/models to geospatial component
