+++ app/models/__init__.py
@@ -1,5 +1,4 @@
 from .changelog import ChangeLog
-from .geospatial_index import GeospatialIndex
 from .location import Location
 from .reportoptions import ReportOption
 from .reports import Report
