+++ app/schemas/__init__.py
@@ -1,6 +1,5 @@
 from .changelog import ChangelogOut
 from .enums import *
-from .geo_index import GeospatialRecordOut, GeospatialRecord, GeospatialRecordCreate
 from .geocoding import *
 from .guest_user import LocationRequestOtp
 from .location import (
